[NS-54][transfer] Re-enable property tests
- Re-enable property tests for bulk transfers as the JVM-crashing bug is now fixed

- Re-write some of the property tests to be more succinct

+++ src/main/resources/com/nec/cyclone/cpp/tests/cyclone_utils_spec.cc
@@ -271,7 +271,7 @@ namespace cyclone::tests {
 
     uint64_t lengths[3] = {16, 56, 59};
 
-    const auto vbytes = sizeof(uint64_t) * frovedis::ceil_div(lengths[0] + lengths[1] + lengths[2], size_t(64));
+    const auto vbytes = sizeof(uint64_t) * frovedis::ceil_div(lengths[0] + lengths[1] + lengths[2], uint64_t(64));
     uint64_t* output = static_cast<uint64_t *>(calloc(vbytes, 1));
 
     cyclone::append_bitsets(output,  0, &inputs[0], lengths[0]);
